backend/requirements.txt
Flask
flask-cors
flask-pymongo
Flask-SQLAlchemy
python-dotenv
PyJWT
werkzeug
